% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_aki.R
\name{detect_aki}
\alias{detect_aki}
\title{Identify Acute Kidney Injury}
\usage{
detect_aki(creatinine)
}
\arguments{
\item{creatinine}{Returns data frame with specific columns:

\strong{ID:} Unique patient identification.

\strong{COLLECTED_DTS:} Dates when corresponding samples were
collected. Must be R datetime object.

\strong{RESULT:} Creatinine value cooresponding to collected sample.

\strong{BIRTH_DTS:} Birth date of patient. Must be R date or
datetime object.

\strong{RI_HIGH:} Reference index for high side of normal expected
creatinine levels.}
}
\value{
Data frame identifying acute kidney injury with specific
colums:

\strong{AKI_FLAG:} TRUE if some AKI event is detected, FALSE if
AKI event was not detected.

\strong{STAGE:} Classifies the stage (1,2,3) of AKI event. Returns
NA if classification is unknown or AKI_FLAG is FALSE.

\strong{BASELINE:} Baseline used in detection of AKI_FLAG. Returns
NA if no baseline was established.
}
\description{
Uses KDIGO Clinical practice guidline for acute kidney injury to
detect AKI events.
}
\details{
Using Kidney Disease: Improving Global Outcomes (KDIGO)
guidelines, Acute Kidney Injury (AKI) are detected using this
function. Guidelines implemented can be found at:

Selby N. M., Hill R., & Fluck R. J. (2015). Standardizing the Early
Identification of Acute Kidney Injury: The NHS England National
Patient Safety Alert. \emph{Nephron, 131, 113-117.}
}
\keyword{kdigo,}
\keyword{kidney}
