% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testset_aki}
\alias{testset_aki}
\title{Test set for detect_aki function.}
\format{A data frame with 50 rows and 4 variables:
\describe{
  \item{ID}{Individual subject identification}
  \item{COLLECTED_DT}{Date creatinine is collected}
  \item{BIRTH_DT}{Birth date of subject}
  \item{RESULT}{Plasma creatinine concentration (mg/dl)}
}}
\usage{
testset_aki
}
\description{
Used to test functionality for detect_aki function.
}
\keyword{datasets}
