% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_aki.R
\name{detect_aki}
\alias{detect_aki}
\title{Identify Acute Kidney Injury}
\usage{
detect_aki(df)
}
\arguments{
\item{df}{Data frame containing specific columns:

\strong{ID:} Unique patient identification.

\strong{COLLECTED_DT:} Dates when corresponding samples were
collected. Must be R datetime object.

\strong{RESULT:} Creatinine value cooresponding to collected sample.

\strong{BIRTH_DT:} Birth date of patient. Must be R date or
datetime object.

Optional:

\strong{ULRI:} Reference index for upper limit of normal expected
creatinine levels. Only required for ages under 18. See note in details
for default value}
}
\value{
Returns a data frame identifying acute kidney injury with specific
colums:

\strong{AKI_STAGE:} Classifies the stage (1,2,3) of AKI event. 0 if there
is no AKI event and NA if classification is unknown.

\strong{AKI_BASELINE:} Baseline used in detection of AKI_FLAG. Returns
NA if no baseline was established.
}
\description{
Uses KDIGO Clinical practice guidline for acute kidney injury to
detect AKI events.
}
\details{
Using Kidney Disease: Improving Global Outcomes (KDIGO)
guidelines, Acute Kidney Injury (AKI) are detected using this
function. Guidelines implemented can be found at[1].

Note: if no reference index (RI_HIGH) is provided for a creatinine
value, a value will be generated using reference values found in
[2]. For ages under 18, the fractional polynomial value will be
used. This value may be used to categorize stage 3 AKI.

1. Selby NM, Hill R, Fluck RJ. Standardizing the early
identification of acute kidney injury: the NHS England ational
patient safety alert. \emph{Nephron.} 2015;131:113-117.

2. Ceriotti F, Boyd JC, Klein G, et al. Reference intervals for
serum creatinine concentrations: assessment of available data
for global application. \emph{Clinical Chemistry.}
2008;54(3):559-566.
}
\keyword{kdigo,}
\keyword{kidney}
